refactor to simplify and add more oop structure
