+++ js/components/index.js
@@ -1,14 +1,15 @@
-export { default as MekaBoards } from './MekaBoards.js';
+export { default as MekaBoards } from './popup/MekaBoards.js';
 export { default as MekaLoader } from './MekaLoader.js';
-export { default as MekaMiniLoader } from './MekaMiniLoader.js';
-export { default as MekaNav } from './MekaNav.js';
-export { default as MekaOptions } from './MekaOptions.js';
-export { default as MekaTasks } from './MekaTasks.js';
-export { default as MekaPopup } from './MekaPopup.js';
-export { default as MekaOption } from './MekaOption.js';
-export { default as MekaOptionFilters } from './MekaOptionFilters.js';
+export { default as MekaMiniLoader } from './settings/MekaMiniLoader.js';
+export { default as MekaNav } from './popup/MekaNav.js';
+export { default as MekaOptions } from './settings/MekaOptions.js';
+export { default as MekaTasks } from './popup/MekaTasks.js';
+export { default as MekaPopup } from './popup/MekaPopup.js';
+export { default as MekaOption } from './settings/MekaOption.js';
+export { default as MekaOptionFilters } from './settings/MekaOptionFilters.js';
 export { default as MekaError } from './MekaError.js';
-export { default as MekaVisualizer} from './MekaVisualizer.js';
-export { default as MekaVisualizerSprints } from './MekaVisualizerSprints.js';
-export { default as MekaVisualizerVariants } from './MekaVisualizerVariants.js';
-
+export { default as MekaVisualizer} from './visualizer/MekaVisualizer.js';
+export { default as MekaVisualizerSprints } from './visualizer/MekaVisualizerSprints.js';
+export { default as MekaVisualizerVariants } from './visualizer/MekaVisualizerVariants.js';
+export { default as VisualizerVariant } from './visualizer/VisualizerVariant.js';
+export { default as MekaPonent} from './MekaPonent.js';
