+++ js/components/visualizer/MekaVisualizerVariants.js
@@ -1,4 +1,4 @@
-import MekaPonent from './MekaPonent.js';
+import MekaPonent from '../MekaPonent.js';
 
 class MekaVisualizerVariants extends MekaPonent {
     constructor() {
